change else to elif

+++ src/scripts/configure_npm_for_heroku.sh
@@ -67,7 +67,7 @@ heroku_prebuild_exists=$(grep -o '"heroku-prebuild":' package.json)
 if [ -n "$scripts_block_exists" ] && [ -n "$heroku_prebuild_exists" ]; then
     echo "heroku-prebuild script already exists. Exiting."
     exit 1
-else [ -n "$scripts_block_exists" ]; then
+elif [ -n "$scripts_block_exists" ]; then
     # Add or create heroku-prebuild to existing scripts block
     echo "Adding heroku-prebuild to scripts."
     jq '.scripts["heroku-prebuild"] = "npm config set $NPM_REGISTRY ; npm config set $NPM_TOKEN"' package.json > package.tmp && mv package.tmp package.json
